Refactor services, entities, repositories (4)

+++ src/main/java/ua/alexkras/hotel/controller/AdminController.java
@@ -60,7 +60,7 @@ public class AdminController {
 
     @GetMapping("/reservation/{id}/select/{apartmentId}")
     public String confirmReservationPage(@PathVariable("id") Integer reservationId,
-                                         @PathVariable("apartmentId") Integer apartmentId,
+                                         @PathVariable("apartmentId") Long apartmentId,
                                          Model model){
 
         Reservation reservation = reservationService.findById(reservationId)
